Add files via upload
This is the main robot object with a few simple methods already added including moving distances and facing markers and walls. It also includes code for finding the start corner and methods to be added for rotating the robot in degrees along with a method for interfacing with the LEDs. I will have a full write-up of the documentation everyone needs to learn by our session tomorrow.