Update robotObjectSim.py
Optimizations made for the face closest token allow for the direction that the robot turns in for facing the marker based on the angle to be changed and recursively calls the function if no tokens are seen with increasingly larger ranges. Further optimizations have been made to prevent wall markers from being used in any part of this program.